backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pillow==10.4.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
python-multipart==0.0.9
tqdm==4.66.5
